Reviewed-on: https://gitea.dessia.tech/DessiA/dessia-common/pulls/98

+++ compile.py
@@ -44,11 +44,10 @@ class ClientWheelDist(wheel.bdist_wheel.bdist_wheel):
         ('exp-year=', None, 'Year of license expiration'),
         ('exp-month=', None, 'Month of year of license expiration'),
         ('exp-day=', None, 'Day of month of license expiration'),
-        ('formats=', None,
-         "formats for source distribution (comma-separated list)"),
         ('getnodes=', None, 'UUID given by getnode (comma-separated list)'),
         ('macs=', None, 'MACS of machine (comma-separated list)'),
-        ('detect-macs', None, 'using this machine macs')
+        ('detect-macs', None, 'using this machine macs'),
+        ('delete-pyx=', None, 'delete compilation files')
         ]
     
     addresses_determination_lines = ['import netifaces\n',
@@ -90,6 +89,7 @@ class ClientWheelDist(wheel.bdist_wheel.bdist_wheel):
         self.macs = None
         self.detect_macs = None
         self.dist_dir = 'client_dist'
+        self.delete_pyx = True
 
     def finalize_options(self):
         """Post-process options."""
@@ -101,7 +101,13 @@ class ClientWheelDist(wheel.bdist_wheel.bdist_wheel):
             
             macs = []
             if self.macs is not None:
+                self.macs = self.macs.replace('[', '')
+                self.macs = self.macs.replace(']', '')
+                self.macs = self.macs.replace(':', '')
+
                 for mac in self.macs.split(','):
+                    if len(mac) != 12:
+                        raise ValueError('A mac address must be 12 digits long, got: {} instead'.format(mac))
                     macs.append(mac)
                 self.macs = macs        
                 print('\nCompiling for macs: {}'.format(self.macs))
@@ -148,6 +154,8 @@ class ClientWheelDist(wheel.bdist_wheel.bdist_wheel):
                 raise NotImplementedError('Unsupported file type: {}'.format(s))
             formats.append(s)
         self.formats = formats
+        
+        self.delete_pyx = self.delete_pyx == 'True'
 
         
     def protection_lines(self):
@@ -389,7 +397,9 @@ class ClientWheelDist(wheel.bdist_wheel.bdist_wheel):
               '{}.{}'.format(*sys.version_info[:2]),  # like 3.7
               wheel_path))
 
-        self.delete_compilation_files()
+        if self.delete_pyx:
+            self.delete_compilation_files()
+            
         if not self.keep_temp:
             logger.info('removing %s', self.bdist_dir)
             if not self.dry_run:
@@ -478,25 +488,23 @@ def get_version():
     # print('version', version)
     return version
 
-
-setup(
-    name = 'dessia_common',
-    version=get_version(),
-    description="Common tools for DessIA software",
-    long_description='',
-    keywords='',
-    url='',
-    author='Steven Masfaraud',
-    author_email='masfaraud@dessia.tech',
-    packages=['dessia_common'],
-    install_requires=['typeguard', 'networkx', 'numpy', 'pandas', 'jinja2',
-                      'mypy_extensions', 'scipy', 'pyDOE',
-                      'netifaces'],
-    python_requires='>=3.7',
-    cmdclass = {'build_ext': build_ext,
-                'cdist_wheel': ClientWheelDist},
-
-    ext_modules = ext_modules,
+## WARNING: THIS must follow setup.py !!!
+setup(name='dessia_common',
+      version=get_version(),
+      description="Common tools for DessIA software",
+      long_description=readme(),
+      keywords=['Dessia', 'SDK', 'engineering'],
+      url='https://github.com/Dessia-tech/dessia-common',
+      author='Steven Masfaraud',
+      author_email='masfaraud@dessia.tech',
+      packages=['dessia_common'],
+      install_requires=['typeguard', 'networkx', 'numpy', 'pandas',
+                        'jinja2==2.11.1',
+                        'mypy_extensions', 'scipy', 'pyDOE'],
+      python_requires='>=3.7',
+      cmdclass = {'build_ext': build_ext,
+                  'cdist_wheel': ClientWheelDist},
+      ext_modules = ext_modules,
 
 )
